Kompletni refactoring s upravou ThesisRepository

+++ src/main/java/cz/newtworks/FGUApplication/controller/ThesisController.java
@@ -33,7 +33,7 @@ public class ThesisController {
     }
 
     @GetMapping("/count")
-    public ResponseEntity<Integer> getThesesCount() {
+    public ResponseEntity<Long> getThesesCount() {
         return ResponseEntity.ok(thesisService.countTheses());
     }
